notebooks/pipeline/requirements.txt
attrs==23.1.0
boto3==1.28.59
botocore==1.31.59
greenlet==3.0.0
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
numpy==1.26.0
pandas==2.1.1
psycopg2-binary==2.9.9
pyarrow==13.0.0
pyproject-toml==0.0.10
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.3
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.21
toml==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
